feat: New program page paths
