+++ lib/graphcms.js
@@ -59,6 +59,7 @@ const getProgramsPaths = async (args) =>
 const getSampleProgramPage = async (args, preview) =>
   await graphCmsClient.request(SampleProgramPageQuery, {
     stage: preview ? 'DRAFT' : 'PUBLISHED',
+    free: true,
     ...args
   })
